// packages/client/src/apollo/types/AllMainCategories.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: AllMainCategories
// ====================================================

export interface AllMainCategories_findAllMainCategories {
  __typename: "MainCategory";
  id: string;
  identifier: string;
  name: string;
}

export interface AllMainCategories {
  findAllMainCategories: AllMainCategories_findAllMainCategories[];
}